Rename routes and controllers according to API spec
