+++ src/api/index.js
@@ -1,10 +1,10 @@
 const express = require("express");
 
 const { healthRouter } = require('../routes/health/health.router')
-const { breadRouter } = require('../routes/bread/bread.router')
+const { postsRouter } = require('../routes/posts/posts.router')
 
 const router = express.Router();
 router.use("/health", healthRouter);
-router.use("/bread", breadRouter);
+router.use("/posts", postsRouter);
 
 module.exports = router;
